tests: move AddTool under fixtures/tools, drop whoami skill
- Extract the test-only AddTool into tests/fixtures/tools/add.py so
  the layout mirrors src/rlm/tools/. conftest now imports it from
  fixtures.tools.add (tests/fixtures/ becomes a package).
- Remove the source-env-propagation test and its whoami_src skill
  fixture; keep the ptc suite focused on the two dispatch forms.
- Add one-line docstrings to each test.
