--- tests/fixtures/skills/whoami_src/pyproject.toml
@@ -1,15 +1,0 @@
-[project]
-name = "rlm-skill-whoami-src"
-version = "0.1.0"
-description = "Echo RLM_TOOL_CALL_SOURCE (test skill)."
-requires-python = ">=3.10"
-
-[project.scripts]
-whoami_src = "whoami_src.whoami_src:main"
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build.targets.wheel]
-packages = ["src/whoami_src"]
