--- tests/fixtures/skills/whoami_src/src/whoami_src/__init__.py
@@ -1,3 +1,0 @@
-from .whoami_src import PARAMETERS, main, run
-
-__all__ = ["PARAMETERS", "main", "run"]
